# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Internal JSON API for challenge and ticket workflows.",
        "title": "Mirror Guard Auth Gateway API",
        "contact": {},
        "license": {
            "name": "Apache-2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/challenges": {
            "post": {
                "description": "Creates a challenge record and returns a signed prefix bound to challenge_id, difficulty, and bind matrix.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Configure challenge",
                "parameters": [
                    {
                        "description": "Challenge configuration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChallengeConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success=true, data=api.ChallengeConfigData",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "400": {
                        "description": "success=false, error.code=invalid_request",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/challenges/verify": {
            "post": {
                "description": "Consumes a configured challenge_id and validates nonce+prefix. Successful verification marks the challenge as used.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Verify configured challenge",
                "parameters": [
                    {
                        "description": "Challenge verification payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChallengeVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success=true, data=api.ChallengeVerifyData",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "400": {
                        "description": "success=false, error.code=invalid_request",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "404": {
                        "description": "success=false, error.code=challenge_not_found",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "success=false, error.code=challenge_replayed",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "410": {
                        "description": "success=false, error.code=challenge_expired",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    }
                }
            }
        },
        "/api/tickets": {
            "post": {
                "description": "Issues a multi-use ticket bound to the provided bind matrix.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Issue ticket",
                "parameters": [
                    {
                        "description": "Ticket issue payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TicketIssueRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success=true, data=api.TicketIssueData",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "400": {
                        "description": "success=false, error.code=invalid_request or ticket_exhausted",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tickets/verify": {
            "post": {
                "description": "Verifies a ticket against bind matrix and consumes one use on each successful check.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Verify ticket",
                "parameters": [
                    {
                        "description": "Ticket verify payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TicketVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success=true, data=api.TicketVerifyData",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "400": {
                        "description": "success=false, error.code=invalid_request or ticket_invalid",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "404": {
                        "description": "success=false, error.code=ticket_not_found",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "410": {
                        "description": "success=false, error.code=ticket_expired or ticket_exhausted",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseEnvelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.BindMatrix": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "ua": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "api.ChallengeConfigRequest": {
            "type": "object",
            "properties": {
                "bind_matrix": {
                    "$ref": "#/definitions/api.BindMatrix"
                },
                "challenge_id": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "integer"
                }
            }
        },
        "api.ChallengeVerifyRequest": {
            "type": "object",
            "properties": {
                "challenge_id": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "prefix": {
                    "type": "string"
                }
            }
        },
        "api.EnvelopeError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.ResponseEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/api.EnvelopeError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.TicketIssueRequest": {
            "type": "object",
            "properties": {
                "bind_matrix": {
                    "$ref": "#/definitions/api.BindMatrix"
                },
                "uses": {
                    "type": "integer"
                }
            }
        },
        "api.TicketVerifyRequest": {
            "type": "object",
            "properties": {
                "bind_matrix": {
                    "$ref": "#/definitions/api.BindMatrix"
                },
                "ticket": {
                    "type": "string"
                }
            }
        }
    }
}